backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-mysqldb==2.0.0
python-dotenv==1.1.0
gunicorn==23.0.0
requests==2.32.3
beautifulsoup4==4.12.3
flask-sqlalchemy==3.1.1
mysql-connector-python==9.1.0
